pycpp: Avoid size_t
Relates to https://github.com/py2many/py2many/issues/491

+++ tests/expected/binit.cpp
@@ -7,7 +7,7 @@
 
 inline int bisect_right(std::vector<int>& data, int item) {
   int low = 0;
-  int high = pycpp::to_int(data.size());
+  int high = pycpp::to_int(static_cast<int>(data.size()));
   while (low < high) {
     int middle = pycpp::to_int((low + high) / 2);
     if (item < data[middle]) {
